Augmented matrix class and tests are written, appear to be working.
MatrixGenerator is giong to become a fully-fledged class but with a single purpose -- removed static methods that were being used everywhere and rolled them in to the two matrix classes.

Changes made to the denseMatrix class to support these changes (gained some functions).

Fixed most of the stuff that broke -- the driver code needs work though.